+++ app/Http/Controllers/TestCategoryController.php
@@ -14,13 +14,18 @@ class TestCategoryController extends Controller
     }
         
     public function index(){
+        if (!getOnlineUser()->can('view-examens-categories')) {
+            return back()->with('error', "Vous n'êtes pas autorisé");
+        }
         $testcategories = CategoryTest::with(['tests'])->get();
         // dd($testcategories);
         return view('tests.category.index',compact(['testcategories']));
     }
 
     public function store(Request $request){
-        
+        if (!getOnlineUser()->can('create-examens-categories')) {
+            return back()->with('error', "Vous n'êtes pas autorisé");
+        }
         $data=$this->validate($request, [
             'code' => 'required',
             'name' => 'required |unique:category_tests,name',          
@@ -37,6 +42,9 @@ class TestCategoryController extends Controller
 
 
     public function update(Request $request){
+        if (!getOnlineUser()->can('edit-examens-categories')) {
+            return back()->with('error', "Vous n'êtes pas autorisé");
+        }
         $data=$this->validate($request, [
             'id2' => 'required',
             'code2' => 'required',
@@ -58,6 +66,9 @@ class TestCategoryController extends Controller
 
 
     public function destroy($id){
+        if (!getOnlineUser()->can('delete-examens-categories')) {
+            return back()->with('error', "Vous n'êtes pas autorisé");
+        }
         $categorytest = CategoryTest::find($id)->delete();
 
         if ($categorytest) {
@@ -69,6 +80,9 @@ class TestCategoryController extends Controller
 
 
     public function edit($id){
+        if (!getOnlineUser()->can('edit-examens-categories')) {
+            return back()->with('error', "Vous n'êtes pas autorisé");
+        }
         $data = CategoryTest::find($id);
         return response()->json($data);
     }
